requirements.txt
--extra-index-url https://www.paddlepaddle.org.cn/packages/stable/cpu/
paddlepaddle==3.2.0
fastapi
uvicorn[standard]
httpx
pymupdf
Pillow
paddleocr
python-docx
chromadb
aiofiles
structlog
click
prometheus-client
python-multipart
pydantic-settings
numpy